+++ CHANGELOG.md
@@ -3,6 +3,9 @@ Change Log
 
 ## 2.0.0-TBD (TBD)
 
+### Changed
+- OpenApi generator updated to release [v4.2.3](https://github.com/OpenAPITools/openapi-generator/releases/tag/v4.2.3)
+
 ## 2.0.0-alpha02 (2019-01-21)
 
 ### Added